+++ src/module/test/testgetdata_lxml/test_officialprogram.py
@@ -7,7 +7,7 @@ getdataモジュール用単体テスト
 import pytest
 
 from module.getdata_lxml import OfficialProgram
-from module.getter import GetParserContent
+from infrastructure.getter import GetParserContent
 
 from ..common import CommonMethod
