+++ src/module/test/testgetdata_lxml/test_officialresults.py
@@ -7,7 +7,7 @@ getdataモジュール用単体テスト
 import pytest
 
 from module.getdata_lxml import OfficialResults
-from module.getter import GetParserContent
+from infrastructure.getter import GetParserContent
 
 from ..common import CommonMethod
